requirement.txt
python-docx
PyMuPDF
PyPDF2
Image
pytesseract